Add pagination component (#195)
Add final draft of Paginator.razor component for paging through a large set of results. Introduces 
error stemming from keeping DbContext class alive in the service layer when working with collections 
of records.

* partial(DAL): provide sample fix for SecurityPriceService select method (#194)

* feat(Blazor): iron out issues with Paginator component

Resolves #185
