+++ NjordinSight.DAL/EntityModelService/Abstractions/ModelReaderService.cs
@@ -153,15 +153,15 @@ namespace NjordinSight.EntityModelService.Abstractions
         {
             int limitPageSize = Clamp(pageSize, 0, 100);
 
-            if (ParentExpression is not null)
-                predicate = ParentExpression.AndAlso(predicate);
+                if (ParentExpression is not null)
+                    predicate = ParentExpression.AndAlso(predicate);
 
-            if (HasSharedContext)
-                return await ReadAsync(SharedContext, predicate, pageNumber, pageSize: limitPageSize);
+                if (HasSharedContext)
+                    return await ReadAsync(SharedContext, predicate, pageNumber, pageSize: limitPageSize);
 
-            using var context = await _contextFactory.CreateDbContextAsync();
+                using var context = await _contextFactory.CreateDbContextAsync();
 
-            return await ReadAsync(context, predicate, pageNumber, pageSize: limitPageSize);
+                return await ReadAsync(context, predicate, pageNumber, pageSize: limitPageSize);
         }
 
         static int Clamp(int value, int min, int max)
