+++ NjordinSight.Web/Components/Common/ContentTile.razor
@@ -1,4 +1,4 @@
-﻿<p class="@Class">
+﻿<div class="@Class">
     @if(RenderTitleSection)
     {
         <div class="content-header">
@@ -22,7 +22,7 @@
             @Body
         </div>
     }
-</p>
+</div>
 
 @code {
     [Parameter]
