+++ NjordinSight.Web/Components/Common/Paginator.razor.css
@@ -1,24 +1,29 @@
-﻿.control-container {
-    display: flex;
-    justify-content: flex-start
-}
-
-.control-container > .form-control:not(button) {
-    /*margin-left: 0.5rem !important;*/
-    width: 100px;
-}
-
-.control-container .pagination-caption {
-    padding-left: 1rem;
-    padding-right: 1rem;
-    font-size: 0.85rem;
-
-    /* This flag does not seem correct. Dev toosl recognize value just fine.*/
-    /* TODO: Investigate. */
-    align-self: last baseline;
-}
-::deep .paginator .icon {
+﻿form.paginator {
+    height: inherit;
     width: 100%;
-    aspect-ratio: 1/1;
-    min-width: 16px;
+}
+
+.control-container {
+    display: flex;
+    justify-content: flex-start;
+}
+
+.control-container > .form-control.button.icon {
+    width: 2rem;
+    background-color: transparent;
+    border: none;
+}
+
+::deep .form-control:not(button) {
+    margin-left: 0.5rem !important;
+    font-size: 0.85rem;
+}
+
+.control-container .paginator-caption {
+    padding: 0rem 0.5rem;
+    font-size: 0.85rem;
+    font-weight: normal;
+    white-space: nowrap;
+    display: flex;
+    align-self: center;
 }
