+++ NjordinSight.Web/Components/DataGrid.razor
@@ -27,11 +27,22 @@
                 }
             </tbody>
             <tfoot class="datagrid-footer">
-                <tr>
-                    <td class="datagrid-cell text" colspan="@Columns.Count()">
-                        @($"{RecordCount} records returned.")
-                    </td>
-                </tr>
+                @if(!SuppressRecordCount)
+                {
+                    <tr>
+                        <td class="datagrid-cell text" colspan="@Columns.Count()">
+                            @Strings.DataGrid_Caption_RecordCount.Format(RecordCount.ToString("N0"))
+                        </td>
+                    </tr>
+                }
+                @if(!string.IsNullOrEmpty(FootNote))
+                {
+                    <tr>
+                        <td class="datagrid-cell text" colspan="@Columns.Count()">
+                            @FootNote
+                        </td>
+                    </tr>
+                }
             </tfoot>
         </table>
     </CascadingValue>
@@ -60,6 +71,18 @@
     public RenderFragment ChildContent { get; set; } = default!;
 
     /// <summary>
+    /// Gets or sets the text foot note for this data grid.
+    /// </summary>
+    [Parameter]
+    public string? FootNote { get; set; }
+
+    /// <summary>
+    /// Gets or sets whether the record count in the grid's data is displayed in the footer.
+    /// </summary>
+    [Parameter]
+    public bool SuppressRecordCount { get; set; }
+
+    /// <summary>
     /// Gets or sets the enumerable collection of <typeparamref name="TModel"/> for this grid.
     /// </summary>
     [Parameter]
@@ -68,7 +91,7 @@
     /// <summary>
     /// Gets the list of <see cref="DataGridColumn{TModel}"/> objects in this grid.
     /// </summary>
-    public IList<DataGridColumn<TModel>> Columns { get; init; } = new List<DataGridColumn<TModel>>();
+    public ICollection<DataGridColumn<TModel>> Columns { get; init; } = new List<DataGridColumn<TModel>>();
 
     private int RecordCount => Data?.Count() ?? default;
