+++ NjordinSight.Web/Components/Strings.resx
@@ -169,11 +169,7 @@
   <data name="Caption_NoContent" xml:space="preserve">
     <value>(No content)</value>
   </data>
-  <data name="Caption_PageCount" xml:space="preserve">
-    <value>{0} pages</value>
-    <comment>{0} is the total count of pages</comment>
-  </data>
-  <data name="Caption_RecordCount_WithTotal" xml:space="preserve">
+  <data name="Caption_RecordCount" xml:space="preserve">
     <value>{0} records of {1}</value>
     <comment>{0} is the count of returned records
 {1} is the total count of records matching the filter</comment>
@@ -191,6 +187,10 @@
   <data name="Caption_ViewSingle" xml:space="preserve">
     <value>View {0}</value>
     <comment>{0} is text indicating the type/record to view, e.g., account or account 1234</comment>
+  </data>
+  <data name="DataGrid_Caption_RecordCount" xml:space="preserve">
+    <value>{0} records returned.</value>
+    <comment>{0} is the record count received from the data store</comment>
   </data>
   <data name="DialogResult_Abort" xml:space="preserve">
     <value>Abort</value>
@@ -235,5 +235,12 @@
   <data name="DialogResult_Yes" xml:space="preserve">
     <value>Yes</value>
     <comment>Indicates the user has affirmatively responded to a question.</comment>
+  </data>
+  <data name="Paginator_Caption_PageCount" xml:space="preserve">
+    <value>of {0} pages</value>
+    <comment>{0} is the total page count</comment>
+  </data>
+  <data name="Paginator_Caption_PageSize" xml:space="preserve">
+    <value>records per page</value>
   </data>
 </root>
